ai-service/requirements.txt
flask==3.0.0
flask-limiter==3.5.0
flask-cors==4.0.0
flask-talisman==1.1.0
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.2
chromadb==0.4.22
sentence-transformers==2.3.1
pytest==7.4.0